requirements.txt
paho.mqtt
python-dotenv
pillow
onnxruntime
ultralytics